+++ ReflecAnnoHW/src/main/java/ru/otus/Main.java
@@ -1,7 +1,7 @@
-package org.example;
+package ru.otus;
 
-import org.example.testRunner.testAnnotations.Exceptions.InvalidTestAnnotationUsageException;
-import org.example.testRunner.TestRunner;
+import ru.otus.testRunner.testAnnotations.Exceptions.InvalidTestAnnotationUsageException;
+import ru.otus.testRunner.TestRunner;
 
 // Press Shift twice to open the Search Everywhere dialog and type `show whitespaces`,
 // then press Enter. You can now see whitespace characters in your code.
