+++ ReflecAnnoHW/src/main/java/ru/otus/testRunner/testAnnotations/AfterSuite.java
@@ -1,4 +1,4 @@
-package org.example.testRunner.testAnnotations;
+package ru.otus.testRunner.testAnnotations;
 
 import java.lang.annotation.ElementType;
 import java.lang.annotation.Retention;
@@ -7,6 +7,5 @@ import java.lang.annotation.Target;
 
 @Retention(RetentionPolicy.RUNTIME)
 @Target(ElementType.METHOD)
-public @interface Test {
-    int priority();
+public @interface AfterSuite {
 }
