+++ src/main/java/ru/otus/Main.java
@@ -1,4 +1,4 @@
-package org.example;
+package ru.otus;
 
 // Press Shift twice to open the Search Everywhere dialog and type `show whitespaces`,
 // then press Enter. You can now see whitespace characters in your code.
